DX-1071: Rename jekyll_commander* files to jekyll_builder*
