+++ .docker/entrypoint/spec/includes.rb
@@ -18,7 +18,7 @@ require 'gemfile_differ'
 require 'gemfile_generator_exec'
 require 'gemfile_generator'
 require 'helpers/spec_jekyll_build'
-require 'helpers/spec_jekyll_commander'
+require 'helpers/spec_jekyll_builder'
 require 'helpers/spec_logger'
 require 'helpers/spec_verifier'
 require 'jekyll_config_provider'
