+++ .docker/entrypoint/spec/jekyll_builder_spec.rb
